pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nigpig"
version = "2.0.0"
description = "NigPig Tools - Security Testing Automation with OWASP ZAP"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "NigPig Team" }
]
keywords = ["security", "testing", "owasp", "zap", "vulnerability"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "python-owasp-zap-v2.4>=0.0.21",
    "aiosqlite>=0.19.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    # Advanced features
    "dnspython>=2.4.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "cryptography>=41.0.0",
    "aiofiles>=23.0.0",
    # Web dashboard
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sse-starlette>=1.8.0",
    # Desktop GUI
    "customtkinter>=5.2.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "types-PyYAML>=6.0.0",
    "types-python-dateutil>=2.8.0",
    "types-beautifulsoup4>=4.12.0",
]
sast = [
    "semgrep>=1.0.0",
]
audit = [
    "pip-audit>=2.6.0",
]
visual = [
    "playwright>=1.40.0",
]
full = [
    "semgrep>=1.0.0",
    "pip-audit>=2.6.0",
    "playwright>=1.40.0",
]

[project.scripts]
nigpig = "nigpig.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["nigpig"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
